# backend/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API de Transacciones Tenpistas
  description: |
    Microservicio para el registro y consulta de transacciones financieras.
    Incluye validaciones de negocio: montos no negativos y fechas no futuras.
  version: 1.0.0

servers:
  - url: http://localhost:8080/v1/api
    description: Servidor local de desarrollo (Spring Boot)

tags:
  - name: Auth
    description: Autenticación de usuarios para acceso al ledger.
  - name: Transactions
    description: Registro y visualización de movimientos de Tenpistas.

paths:
  # --- AUTENTICACIÓN ---
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login de usuario
      description: Autentica al usuario y retorna el token JWT para operar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # --- TRANSACCIONES ---
  /transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: Listar transacciones
      description: Retorna el historial de transacciones paginado y ordenado por fecha descendente.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: tenpistaName
          in: query
          description: Filtro opcional por nombre del Tenpista (búsqueda parcial, case-insensitive)
          required: false
          schema:
            type: string
        - name: commerceName
          in: query
          description: Filtro opcional por comercio/giro (búsqueda parcial, case-insensitive)
          required: false
          schema:
            type: string
        - name: fromDate
          in: query
          description: Fecha/hora inicial (inclusive) para filtrar por transaction_date
          required: false
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          description: Fecha/hora final (inclusive) para filtrar por transaction_date
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Número de página (base 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Tamaño de página
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Listado obtenido correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Crear nueva transacción
      description: Registra un movimiento. El monto debe ser > 0 y la fecha <= actual.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '201':
          description: Transacción registrada con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Error en validaciones (ej. monto negativo o fecha futura).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/BadRequest' }
    Unauthorized:
      description: Token inválido o inexistente.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Unauthorized' }
    InternalServerError:
      description: Error crítico en el servidor.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/InternalServerError' }

  schemas:
    # --- BASE DE ERROR ---
    ErrorResponse:
      type: object
      properties:
        code:      { type: integer, format: int32 }
        message:   { type: string }
        timestamp: { type: string, format: date-time }
        details:   { type: string }

    BadRequest:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - example:
            code: 400
            message: "Validación Fallida"
            timestamp: "2026-04-09T13:20:00Z"
            details: "El monto no puede ser negativo."

    Unauthorized:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - example:
            code: 401
            message: "No autorizado"
            timestamp: "2026-04-09T13:20:00Z"
            details: "Token JWT expirado o malformado."

    InternalServerError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'

    # --- AUTENTICACIÓN ---
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: "admin@tenpo.cl" }
        password: { type: string, format: password, example: "Tenpista2026!" }

    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type:   { type: string, example: "Bearer" }

    # --- TRANSACCIONES ---
    TransactionCreateRequest:
      type: object
      required: [amount, commerce_name, tenpista_name, transaction_date]
      properties:
        amount:
          type: integer
          minimum: 1
          description: "Monto en pesos (debe ser mayor a 0)"
          example: 5000
        commerce_name:
          type: string
          description: "Giro o comercio de la transacción"
          example: "Starbucks"
        tenpista_name:
          type: string
          description: "Nombre del Tenpista"
          example: "Cristian"
        transaction_date:
          type: string
          format: date-time
          description: "Fecha y hora de la transacción (no puede ser futura)"
          example: "2026-04-09T10:30:00Z"

    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 101
        amount:           { type: integer, example: 5000 }
        commerce_name:    { type: string,  example: "Starbucks" }
        tenpista_name:    { type: string,  example: "Cristian" }
        transaction_date: { type: string,  format: date-time }
        created_at:       { type: string,  format: date-time }

    TransactionPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        total_elements:
          type: integer
          format: int64
          description: "Total de transacciones"
          example: 100
        total_pages:
          type: integer
          description: "Total de páginas"
          example: 10
        page:
          type: integer
          description: "Página actual (base 0)"
          example: 0
        size:
          type: integer
          description: "Tamaño de página"
          example: 10